chore(dapp): rename prop and relocate NamePurchaseCard component for improved structure
